Move glyphs to their own file
